requirements.txt
lxml
Pillow
requests
Flask
uwsgi